// IMFS-Web-FrontEnd/src/app/models/application/application.model.ts
export class ApplicationContact {
    contactType: number;
    contactDescription: string;
    contactID: string;
    contactEmail: string;
    resellerID: string;
    contactName: string;
    contactDOB?: Date;
    contactAddress: string;
    contactDriversLicNo: string;
    contactABNACN: string;
    contactPosition: string;
    isContactSignatory?: boolean;
    contactPhone: string;
}